# std/crypto: make Gimli 60% faster