Before:

           gimli-hash:        120 MiB/s
           gimli-aead:        130 MiB/s

After:

           gimli-hash:        195 MiB/s
           gimli-aead:        208 MiB/s


 Also fixes in-place decryption by the way, so this supersedes #6449