@jedisct1 2020-09-28
Fixed, same performance improvement :)