@jedisct1 2020-09-29
Changed, thanks!

`permute_small` can be renamed, but we need a name that doesn't collide with the `permute` constant :)